# xor_keystream(): Use unsafe, and process one byte at a time.