@DaGenix 2015-02-06
I played around with trying to use `usize`s or `u32`s to do the XORs if the buffers are properly aligned, but that seemed to only hurt performance, at least for the small-ish buffers that all of the current algorithms use.

Thanks!